wayland/subsurfaces: Move wl_subsurface fields into MetaWaylandSuburface
In order to isolate subsurface related logic to
meta-wayland-subsurface.c, move all the fields previously kept in
MetaWaylandSurface into MetaWaylandSubsurface - the role instance.

This also makes the user data pointer of the subsurface resource point
to the MetaWaylandSubsurface object instance, as during surface
destruction, it would otherwise not be possible fetch the subsurface
instance via the role pointer.

https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/1342