wayland/surface: Introduce surface state fences
A fence in this context is a way to control when state is actually
applied to surface. It works by having an priority ordered list of fence
callbacks that at the time of wl_surface_commit() either consumes a
pending state, or lets it being processed by the next fence in the list.

It then lets fences that consumed a pending state to postpone the
application of the said state until a later point in time. When it wants
to do this, it will commit the state via the state fences but only let
fences with a lower priority than itself process it.

The API works by having fence owners add a fence, with a priority. On
commit, the fence callback returns TRUE if it consumed the state, and
FALSE if it did not. Priorities are declared at a MetaWaylandSurface
level, but each priority only has a single owner. The lower priority,
the higher priority number.

The 'base' priority is the last step, with lowest priority, where the
state is finally applied to the surface.

The 'top' priority is a dummy priority with the highest value; it is ony
used when intending to process all fences.

This commit both introduces and ports subsurfaces to use the fencing
instead of manual cached state management in MetaWaylandSurface.

The motivation for this change is to allow for multiple levels of state
synchronization management.

https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/1342