renderer-view/native: Allow syncing frame clock updates with actor updates
Allow per-view tracking of actors for the purpose of synchronizing
the frame clock to their updates.

This uses the variable mode of ClutterFrameClock to either perform an
immediate update or a delayed update depending on whether the update
is performed in the context of the tracked actor or a different one.

This approach achieves a minimal latency between updates of the actor
and updates of the view, and delays updates of the view when the
tracked actor has not updated.

Changes in the synchronization mode are applied asynchronously at the
preparation phase of the next frame.

If the tracked actor is frozen or destroyed, the synchornization is
stopped as no further updates from the actor are expected.

For now, the only applicable synchronization mode is
META_FRAME_SYNC_MODE_DISABLED, which doesn't synchronize to actor
updates and keeps the frame clock on the its default fixed mode.
When support for VRR is added in the commits to follow,
META_FRAME_SYNC_MODE_ENABLED will be set when appropriate.