clutter/content: Add API to query opaque region of ClutterContents
We're adding obscured region tracking to Clutter, to make this actually
useful, we need to allow ClutterContent implementations to provide an
opaque region for us. So do that and add a new get_opaque_region()
vfunc to ClutterContent, implementations of this vfunc should hand over
a reference to a cairo_region_t that includes the opaque region in
coordinates relative to the texture.

Also add a function clutter_content_invalidate_opaque_region() that can
be called in case the opaque region returned by the get_opaque_region()
implementation has changed.