clutter/util: Add function to convert GrapheneRect to integers
Add a function to clutter-util which converts GrapheneRects to a
cairo_rectangle_int_t, returning the largest integer rectangle that's
contained by the floating point rectangle.

This function tries to adjust for inaccuracies from floating point
calculations by using rounding the float to the nearest integer in case
that integer is within a 0.001 range of the original value. Doing this
is very important because we're going to use this function for creating
cairo regions that are used for culling, and we want to be able to
reliably cull out actors which are aligned to the pixel grid (like
windows).