clutter: Add obscured region tracking to ClutterActor
Start building the structure necessary to implement obscured region
culling in Clutter: The important part here is a function called
clutter_actor_build_unobscured_region() which loops through the actor
tree from the topmost visible actor down to the bottommost one. This
recursive process builds the unobscured regions for every visible actor
in the actor tree: We start with a region that has the full size of the
stage and we then subtract every actors opaque region, resulting in the
unobscured region for the actor underneath.

clutter: Store less copies of unobscured regions

Not every actor provides an opaque region to use when creating the
unobscured region, in fact most actors don't. Also most actors don't
have a clip set. This means that most actors don't influence the
unobscured region at all, and storing a separate copy of the same region
for all those actors is a waste of memory.

So optimize the resource usage of the unobscured region machinery by
using references to regions instead of copies when regions are the same
across actors.

We do this by adding a third region to ClutterActor:
priv->unobscured_region. This is a region where not only the opaque
regions of all the actors above are subtracted, but also the one of the
actor itself (that means it's the equivalent of the
unobscured_without_children region of the actor underneath). This new
region is the only one we actively change, the remaining two regions are
from now on simply references to an unobscured_region of a different
actor (or the stage_region of the stage in case of the topmost actor).