clutter/actor: Cull out actors fully obscured by other actors
Make use of the newly added obscured region tracking and extend the
culling to also cull out actors which are completely obscured by other
actors:

- When painting an actor where the redraw clip doesn't intersect with
the unobscured_without_children region, we can completely stop the paint
process for the whole subtree. That's because we know the whole are
being redrawn is fully obscured by an actor on top of us and our
children. We can use the redraw clip for the intersection check here
because the paint-volume culling done before our check would already
have stopped the painting if the actor was outside the redraw clip.

- When painting an actor where the redraw clip doesn't intersect with
the unobscured_including_children region, we know that the area being
redrawn is fully obscured by a child of ourselves, which allows us to
skip painting our own texture. Since the area a child draws to can be
larger than the area of the actors texture, ideally we'd want to
intersect the redraw clip with the actors allocation here, that might be
a future optimization.

Since we don't need the actual intersecting region when performing the
culling, we use cairo_region_check_intersects() to implement the
hot-path checks instead of cairo_region_intersect().