clutter/actor: Add API to get the unobscured region
Certain actors, like MetaSurfaceActor want to know the exact unobscured
region of an actor, for example to figure out whether it's obscured only
one specific stage view.

It seems that a simple clutter_actor_is_obscured() API won't cover all
needs there, so add slightly more sohisticated API that allows getting
the unobscured in both stage and actor coordinates.

The method is made that way with the two region-pointers to allow for
getting both regions with minimal overhead, as for the local region, the
stage region needs to be created beforehand anyway.