clutter/actor: Make clips modify the opaque/unobscured region
When an actor has enabled clip_to_allocation or a clip is manually set,
the rectangle that the actor and its children can draw to gets limited.
That also means the opaque region of the actor and the regions of all
its children have to be limited to the clip area.

Since we don't have the opaque region of the actor including its
children readily available, we get that region by subtracting the
unobscured region as it looked like before our actor from the unobscured
region including our actor and its children.

Also add a test to make sure it all works as expected.