clutter: Introduce SIZE_REQUESTS debug flag
When debugging using the LAYOUT flag, mutter gets quite slow and logs a
lot of debugging messages from the size caching machinery. This is
helpful when the issue is about size requests, but makes the flag almost
unusable in all other cases. So introduce a separate SIZE_REQUESTS
ClutterDebugFlag which can be used to look into size-related issues and
make the LAYOUT flag log messages for everything else.