clutter/actor: Warn if not all mapped children were allocated
Clutter expects implementations of the allocate() vfunc to allocate all
mapped children of an actor, if they skip any actors, bad things will
happen.

So go through the children after calling the allocate() vfunc to check
whether they all have up-to-date allocations and print a warning if
that's not the case, this should catch all the bad citizens. Since
looping through the children can be quite costly, only do this check
when the LAYOUT debug flag is set.