clutter/timeline: Remove markers
We've never used them anywhere and Clutter will see some refactoring
around the animation API which markers would only complicate, so remove
them.

Since markers are the only reason ClutterTimeline implements
ClutterScriptable, we can stop implementing that interface now.

On the testing side of things, the frame-clock-timeline test made use of
a marker to switch the frame clock that's driving the timeline after 500
milliseconds. This was replaced with a "new-frame" handler which uses
the elapsed time to achieve the same thing.