Add empty sdk application
While just being a header bar and an about dialog, the aim of this
application is to provide a more useful interface for running mutter and
gnome-shell as a nested display server, for debugging purposes.

The icon used for now is 'default_window.png', as well as the same image
scaled to 64x64 and 128x128.

In contrast to the meta-*/Meta* prefix, the SDK application has the
mdk-*/Mdk* prefix, standing for Meta/Mutter Development Kit.