screen-cast/area-stream: Use monitor framebuffer in case area is monitor
The largest part of computer users uses only a single monitor, in this
case we can avoid painting the stage a second time in the area stream
and simply use the monitors framebuffer as the source of our recording.

So introduce an optimization for the area stream that compares our
rectangle to the logical monitor rects that exist, and make use of the
functions to record monitor buffers/framebuffers that we factored out of
the monitor stream source in the last commits.