xwayland: Make X11 sockets world writable
The X11 server has its own authentication mechanism, we do not (and
should not) need to rely on UNIX permissions for that, otherwise that
prevents X11 clients from other users from connecting to the Xserver
even though the X11 credentials are valid.

Make sure we create sockets writable to all.

Closes: https://gitlab.gnome.org/GNOME/mutter/-/issues/1468