constraints: Ignore min/max limits for fullscreen windows
Constraining fullscreen windows leads to an inconsistent shell behaviour,
where the shell behaves as if a window is in fullscreen mode (e.g. hides
some shell UI elements), while the actual application window will only
cover a subarea of the entire screen. By not constraining fullscreen
windows, we allow the application to decide if the entire screen should be
covered or not.