backends/native: Add scrolling speed configuration
This commit implements a less naive approach of configuring scrolling
speed for mice and touchpads. Due to implementation limitations, the
change only applies to the native backend.

To achieve the goal, a GObject property is added to MetaInputDevice
and a struct field is added to ClutterScrollEvent, both called
"scroll-speed". Those properties are used as a coefficient when
sending scroll events to Wayland clients. This approach is ignored on
X11. The property gets set for devices via GSettings similar to other
settings.

Fixes #551