clutter/actor: Require a context on construction
An actor is an object that is tied to a Clutter context in one way or
the other. In the past, an actor was constructed without an outside
connection (to a context) and instead relied on finding the "default"
backend or context, depending on the use case.

With this commit, we're starting the migration to explicit-context
construction of all Clutter objects that need it, much similar to what
the Cogl 1.0 -> Cogl 2.0 API migration did.

This allows for better control of object lifetimes, and in theory,
multiple Clutter contexts and backends, even though this is unlikely to
be utilized any time soon.