clutter: Introduce ClutterPanGesture for recognizing pan gestures
Replace the existing ClutterPanAction with a new ClutterGesture subclass
for recognizing pan gestures. It works a bit differently from the old
gesture and, like all new ClutterGestures, tries to keep a clear separation
between recognizing the gesture and actually changing state of the
actor. The latter usually requires extra context that the gesture
doesn't have, which is why it's left to the user of the gesture.