renderer-view/native: Enable VRR and actor sync when applicable
If an actor is tracked for synchronization and VRR is configured to
be enabled for the output, set the synchronization mode to
META_FRAME_SYNC_MODE_ENABLED.

If an actor isn't tracked for synchronization or VRR is not configured
to be enabled for the output, set the synchronization mode to
META_FRAME_SYNC_MODE_DISABLED.

When the synchronization mode is set to META_FRAME_SYNC_MODE_ENABLED,
enable the VRR mode on the output, and when it's set to
META_FRAME_SYNC_MODE_DISABLED, disable it.

Enabling the VRR mode on an output along with synchronizing frame
clock updates to actor updates on the matching view allows the monitor
match its refresh rate to the rate at which the actor updates,
effectively achieving the VRR effect.