backend/native: support for three finger drag
After initially looking at implementing this feature in libinput itself,
it was determined that it made more sense for the compositing layer to
emulate the functionality.

This patch adds the necessary emulation for the native/Wayland input
handling backend in Mutter.

Closes: https://bugzilla.gnome.org/show_bug.cgi?id=768421
Supercedes: https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/1021

--

Note that this will only work on Wayland, the X11 implementation is left empty intentionally.