wayland/data-device: Prefer copy when shift+control used in dnd
X11 provides the ability to create a link by holding CTRL+SHIFT
during a DnD operation.  In wayland, the modifiers CTRL and SHIFT
are checked independently to either use the MOVE or COPY action.
Given that COPY is non-destructive, prefer COPY over MOVE
when both control and shift are used.

Fixes: https://gitlab.gnome.org/GNOME/mutter/-/issues/2380