window: Add support for vertically split tiled windows
We supported horizontally tiled windows in horizontally disposed
monitors, but we had no similar implementation when they are disposed
vertically.
And in such setup it's often required to tile windows in cascading form.

So adapt the current logic that we can handle the top-bottom tiling
case.

Keybindings are <Super><Ctrl>Up/Down, and as per this also add the
<Super><Ctrl>Left/Right aliases.

In case edge-tiling is enabled we also support creating top-splits but
only if we have another bottom-split defined, so tiling to the bottom
is required to create top-tiles too.
In this way we can both support the edge-gesture for both maximization
and top-splitting without confusion.

When an horizontal / vertical split is defined in a workarea, then
only splits of the same type (horizontal or vertical) are allowed.
This is because, not having tiling groups defined yet, allowing both
cases would create confusion and partially covered windows.

Fixes: #1596