clutter/actor: Update visible paint volume while getting redraw clip
With the last commit we introduced caching and invalidation logic for
the visible paint volume, that already improved things a lot, but we can
do even better:

For getting a redraw clip, we need both the extents of the actor that
are currently on the screen, as well as the extents of the actor that
will be on screen with the next redraw. That's because in case the actor
gets moved with the redraw, the old texture is also painted over. For
this, so far we just passed visible_paint_volume as the old paint
volume, and the new paint volume we fetched from scratch, forcing the
stage to tranform it to eye coordinates.

There's a smart trick to avoid those extra transforms: Getting the
redraw clip happens before the finish-layout
(clutter_actor_finish_layout()) step in stage updates. This means in
clutter_actor_get_redraw_clip() we'll see
priv->needs_visible_paint_volume_update is TRUE, and can then update the
visible paint volume early. This avoids the extra update in
clutter_actor_finish_layout() afterwards, and we're a little faster
again.