clutter/actor: Use visible_paint_volume for updating stage views list
We have the visible_paint_volume around already and it is in
eye-coordinates, so this saves a bit of resources otherwise used for
matrix multiplications.

Also it's more correct to figure out the stage views we're on by looking
at the paint volume instead of the allocation, since that includes
transforms like scale_x/y or transform_x/y.