clutter/stage: Always queue events in _clutter_stage_queue_event()
The naming of this function is fairly clear, and not queuing an event
when calling _clutter_stage_queue_event() might be a bit unexpected.

The current case where this function doesn't queue an event is if it's
the first event and that is non compressible. This seems to be a
(somewhat unnecessary) optimization that will only make our lives harder
when we actually expect the event to be queued (eg. in tests), so remove
that.