window-actor/wayland: Handle too large fullscreen surfaces better
Some clients, such as kitty, can end up with fullscreen surfaces that
exceed the size limit set by the compositor, i.e. larger than the screen
resolution. Mutter used to place the top left corner of the window
geometry of such surfaces at the top left corner of the screen and these
clients relied on that. This however got changed as a side effect of
909616b2, which now causes these surfaces to be shifted to far to the
top left.

While this client behavior goes against the spec, it is easy enough to
work around in Mutter and restore the previous behavior for such
fullscreen surfaces that exceed the screen size.

Closes: https://gitlab.gnome.org/GNOME/mutter/-/issues/2524