meta: Option to hide pointer when typing
The ability to hide the mouse pointer when typing has been missing
from the Wayland compositor. It was not needed in X11 as there are 3rd
party utilities that could do this, but in Wayland this must be done
by the compositor.

This change introduces a new setting and the ability to hide the
pointer when typing occurrs.

Typing is considered to be keys that aren't modifiers (ctrl, alt, etc)
and also not navigation (page up, etc).

Closes: https://gitlab.gnome.org/GNOME/mutter/-/issues/1865