xwayland: Prepare XAUTHORITY as early as possible
Xauth is communicated to the X11 display connection only via the
$XAUTHORITY environment variable. Setting an environment variable when
there are multiple threads may lead to crashes, so to avoid doing that
rather late when starting up, when there likely are multiple threads
already, do so as one of the first things we ever do.

What is prepared up front in MetaContext is only the file name and file
descriptor itself, as well as the environment variable, the rest is
still handled where it was before by Xwayland.