x11/frame: Always wrap client window in a wrapper window
The wrapper window aims to result in no extra blits in the X server by
using an identical visual and size as the parent window, and should
effectively not be visible ignoring the racyness of X11 window resizes.

The aim for this is to later re-introduce unparenting a window from the
frame window to avoid extra blits when fullscreen. Unparenting the
client window themselves triggers unfixable bugs in old X11 clients that
never expect to change, but unparenting the wrapper instead avoids these
issues.