frame: Delete NEEDS_FRAME property even if we didn't get one
This fixes a race condition where we'd request a frame, then unrequest
it before we got the first frame we requeted. In that case, we'd miss
removing the property, which means the frames client would have it
lingering. The end result is that the next time we'd request one, we
wouldn't get a new frame window, since the frames client ignored the
request, thinking the old one still counts.