test/runner: Synchronize with frames client when waiting
When metatest do 'wait', it is expected that all actions have completed
in every place. This was not true for actions that involved the frames
client, as we could have sent a request (e.g. changed a window
property), round tripped to the X server, and round tripped all the test
clients themselves. What we hadn't done is ensure that all the requests
we expected to have the frames client handled had been seen by the X
server.

To achieve this, introduce another root window property called
`_MUTTER_FRAME_SYNC_SERIALS`. It's a two sized XA_CARDINAL (two
numbers). Whenever the test runner needs to "wait", it'll set the first
of the two to an ever increasing counter, and wait for the second value
to have been updated to match the first.

The frames client is changed to listen on property changes on the same
atom, and will take whatever it sees in the first number, and write it
to the second.

This way we can be sure property changes we sent to the frames client
have been fully processed.

This fixes flakyness of the `closed-transient-only-take-focus-parents`
test.