core: Do not select SubstructureNotify on both frame/wrapper windows
With the introduction of an extra window in the frame hierarchy, and the
selection for Structure and Substructure notification event masks in both
frame and wrapper windows, we receive an excess of ReparentNotify events
for every operation involving window reparenting. One event from the window
being reparented itself (e.g. the wrapper window, through StructureNotifyMask),
and other possible set of events from its old/new parents (depending on the
window being reparented, either of the root, frame, or wrapper window).

We do require SubstructureNotifyMask on the wrapper window in order to receive
DestroyNotify/MapNotify and other desirable events from client windows, and we
do also require it on the root window to track new windows. In order to minimize
the overlap, avoid any of these events on the frame window (already notified
through the root window), and set StructureNotifyMask|SubstructureNotifyMask on
the wrapper window to track changes from there downwards.

This leaves just one excess ReparentNotify maybe triggering
`window->reparents_pending` accounting, which is the one emitted on the root
window after the client window gets reparented to the wrapper window. Ignore
this one for MetaWindow accounting purposes, which makes the accounting nicely
even.