window: Don't unqueue move-resize changes on wayland configuration ACKs
meta_window_wayland_finish_move_resize() may end up being called when a
resize was queued already, in such case we need to perform it promptly
or we'd loose changes that have been requested already.

When repeating multiple times the strut-monitor-changes test this code
was hit consistently, and this change fixes it too.

Co-Authored-By: Marco Trevisan (Treviño) <mail@3v1n0.net>

Closes: GNOME/mutter#1627