backends/native: Add HDR metadata passthrough support
HDR Metadata pass-through for fullscreen surfaces:
When a HDR surface is fullscreen and client presents the HDR content
with metadata and display panel/hw supports HDR, nothing is done by
the KMS i.e. no modifications to the content and metadata is only
passed through from client to KMS display.