tests: Allow most tests to run in parallel
The wrapper is now able to handle this so do it, but still run
xwayland and x11 tests as standalone.

They may use shared resources, so it's safer not to run it in
parallel for now, or it may fail.

Similarly, do not enable parallel mode for the monitor-unit tests
because stressing it, may trigger Gtk bugs in our test clients.
The bugs are handled, but it's better to wait for now.