ci: Use default test setup for testing and no global dbus mocks
We want to use the dbus wrapper for each test, to properly isolate each
one.

In order to run all the tests, we need to also run the ones in the
no-catch suite too.

It may be slower but running in parallel still gives us the same speedup
(or better).