tests: Run native-unit test using wireplumber
It's the only test that actually requires a pipewire session and it's
also a test that is currently not easy to run by distributors or
developers in a clean environment because they have to mimic what we do
in our CI script.

Since we can handle it all in meson, let's move the setup to the test
phase.