wayland/window: Do not apply configurations matching pending ones
During window configurations we may end up in this state:
 - Window is resized to, say 800x600
 - A window is requested to be reconfigured to have a new size of 40x30
 - While the new state is not yet acked a new configuration request to
   resize again the window to 800x600 is performed.
   => This request will be never performed on the XDG toplevel because
      we didn't ack the 20x40 request yet, so the constrained rect
      comparison check fails (given the window is formally still
      considered to be 800x600).

To avoid this to happen we should try compare the new requested
configuration rectangle with the last pending configuration, so that
situations like the one above cannot happen.

Also, to make ops such as Move or resize only to work properly we must
now refer to the pending rectangles values or we'd end up re-configuring
a window wrongly.

As per this, the strut-monitor-changes test is now fixed.

Fixes: https://gitlab.gnome.org/GNOME/mutter/-/issues/1627