window: Partial fix for struts covering maximized windows
Previously, an ACKed configuration from a Wayland window would always
dequeue any pending constrain operations on that window. That
generally is a good idea, but in some cases, the ACK is more stale
than the pending constrain. Specifically, there are situations where
struts change rapdily (screen unlock), and with each change a
constrain in queued which then results in a resize sent to the client
and then a corresponding ACK coming back. If the second constrain is
queued and then the first ACK comes back, the second contrain would
never happen and the user would see struts covering their window.

To solve this, if the server resizes a client window, don't let those
corresponding ACKs dequeue pending constrain operations.