wayland/window: Correctly handle resize from left/top of window
In a move/resize scenario, when deciding whether to sent a configure
message to a client, mutter compares the new size with the old size.
The old size was taken from a pending configure request sent to the
client (if any) or the existing window size (otherwise).

A problem occurs when slow dragging a window on the left or top edge,
though: sub-pixel movements of the input device will cause repeated
resize events to be processed. The first for a given size will be sent
to the client correctly, but then the next will see the pending resize
and consider the resize not a new size and not send to the client.
That's not terrible, but what happens next is that, b/c there is
assumed to not be a resize in this case, this is assumed incorrectly
to be a window move operation. That results in the window moving to
the new left/top.

This commit fixes this problem by forcing all user-initiated resizes
to go through the client.