clutter/actor: Set the top-most actor for stage-views
This is currently a very naive approach: force a full tree traversal
in `clutter_actor_finish_layout()` and always set the current actor as
the topmost one for all stage-views the actor is on. As we traverse the
tree in paint order we'll end up with the right ones set.