compositor: Allow resizing window to the original size
Since the anchor_root_x/y coordinates are fetched at the start
of the resize operation and does not get updated in between,
there may be situations where the resize is mistakenly avoided
as a no-op. E.g. attempts to resize a window, then restoring to
the exact same original size, this may get easier to trigger
with resizes from the bottom/right corner of windows aligned to
stage boundaries.

Closes: https://gitlab.gnome.org/GNOME/mutter/-/issues/2837