compositor: Maybe detach surface actor
As it seems, a quick sequence of hiding and showing a window may
result in the surface actor being incorrectly handled into the
new hierarchy while it's still attached to the old one.

This somehow results in clutter_actor_insert_child_at_index()
since the actor still does have a parent. Can be seen in the
"default-size" stacking test after porting the test client to
GTK4.

Prevent this, and unparent/reparent the actor in that situation.