x11-tests: Wait for zenity clients to be fully terminated
We were just killing them, and this may have lead these children to be
stopped at later point making the whole script to fail.

To ensure this is not the case, let's wait them and ensure that they've
been terminated correctly.

As per this we can enable again catch with x11 tests