backend/x11: Only call keymap_changed for new devices
The x11 backend only calls keymap_changed when the device
triggering a XkbNewKeyboardNotify is first connected.
This prevents it from being called when switching between
multiple keyboard input devices (which can include mouse
buttons). This is a workaround to issue #398, but does
not address any underlying issues with XIGrabKeycode.

Fixes #398