clutter/actor: Add insert_child_before/after methods
The existing insert_child_below/above methods uses a model of
children being organized in a pile from bottom to top.

That can be confusing when applied to actual containers or layout
managers, for example when children are laid out horizontally, or
vertically from top to bottom.

Take a clue from GtkWidget, which uses insert_before/after for
the same functionality, and add copies of the existing functions
that use the less confusing terminology.