kms/impl-device/atomic: Use non-blocking atomic commits for modesets
If there are no page flip listeners, use the OUT_FENCE_PTR CRTC property
to get a sync file descriptor which signals (becomes readable) when the
atomic commit finishes.

In the future, this could be extended to handle all modesets fully
asynchronously, allowing them to be processed in parallel with other
atomic commits by the kernel.

For now though, just wait synchronously for the sync files to signal.
This should result in mostly the same behaviour as far as the rest of
mutter is concerned, and should be enough to avoid hitting the RT CPU
time limit in the kernel.

Issue: https://gitlab.gnome.org/GNOME/mutter/-/issues/3037

v2:
* Use g_clear_fd in process_wait_out_fence (Pascal Nowack)
* Set error as needed in process_wait_out_fence
* Check return value of process_wait_out_fence in
  meta_kms_impl_device_atomic_process_update
v3:
* Get out fence only for enabled CRTCs, to avoid atomic commit ioctl
  failure
* Queue mode set updates if there's a pending page flip in
  meta_kms_impl_device_handle_update
* Get and wait for out fences only if there are no page flip listeners