clutter/frame-clock: Plot dispatch lateness
Allows to see at a glance when something had stalled the main loop
preventing a scheduled frame.

The plot name must be leaked: Tracy plot names must be accessible
throughout the program's lifetime, and our plot name is generated at
runtime to put the output name in (for multi-monitor use-case). Leaking
is the current intended workaround for dynamic frame and plot names:
https://github.com/wolfpld/tracy/issues/608#issuecomment-1705256136