Add Tracy frame instrumentation
Frame instrumentation lets you see at a glance in the Tracy UI which
frames took longer than intended.

We have to use discontinuous Tracy frames to get the most useful data
because Wayland compositors are different from games in two important
aspects:

1. When multiple monitors are connected, they have separate frame sets,
   requiring the use of dynamic frame set names.
2. Wayland compositors don't draw continuously, requiring use of
   carefully positioned discontinuous frame sets.

With this instrumentation, we're mainly interested in finding frame
drops during *continuous* drawing. Because when idling, if the input
event arrives late in a frame, we will naturally only present it one
frame late. This is entirely expected, so we don't want this to pollute
the profiler UI obscuring unintended dropped frames.

Thus, we start a Tracy frame only upon getting a presented event (which
should arrive as soon as possible after a VBlank and *not* while
idling), and then if we detect that we're going to idle, we stop the
frame prematurely.

Ideally we'd start the frame even earlier, right in the page flip
callback, to also catch main loop stalls delaying the presented event
processing. But I'm not sure how to thread it through nicely, ensuring
that the frame is always stopped early if the code doesn't end up in the
presented handler. Either way, the way this commit does it also works
with the nested backend, so that's nice.